cuda with index instead of default "cuda"
In `feature_extractor.py`, images are put to `self.device` every time it's called.
It's necessary to consider this because we should make sure that images(the inputs) and the model parameters are on the same device.
However, the current code cannot achieve this with the default setting. Currently, Feature extractor is initialized with the default `device='cuda'`, it's ambiguous for that "cuda" with no specified index may point to different GPUs.
I modified a line which set `self.cuda` with a specified device index when none is provided.